chore(Refactor): mkae available all error entities
