+++ packages/common-module/src/BussinesError/toBusinessError.js
@@ -1,0 +1,8 @@
+const { BusinessError } = require('./BusinessError')
+
+const toBusinessError = (error, module = 'cvfy') => {
+  if (error instanceof BusinessError) return error
+  return new BusinessError(module, error.message)
+}
+
+module.exports = toBusinessError
